requirements.txt
autopep8==1.6.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
greenlet==1.1.2
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pydantic==1.8.2
python-dotenv==0.19.2
requests==2.26.0
SQLAlchemy==1.4.28
sqlalchemy2-stubs==0.0.2a19
sqlmodel==0.0.5
toml==0.10.2
tomli==1.2.3
typer==0.4.0
typing-extensions==4.0.1
urllib3==1.26.7
